Rename /lib to /utils, move rec schema related functions out of /types
