+++ src/app/layout.tsx
@@ -1,5 +1,5 @@
 import "../styles/globals.css";
-import { cn } from "@/lib/utils";
+import { cn } from "@/utils/utils";
 import { ThemeProvider } from "@/components/theme-provider";
 
 import type { ReactNode } from "react";
