+++ src/db/mongo/schemas/recordedGame/PlayerData.ts
@@ -1,10 +1,15 @@
 import mongoose from "mongoose";
-import { RecordedGamePlayerMetadataKeySchema } from "@/types/RecordedGame";
+import { recMetadataSchemaHelper } from "@/utils/utils";
+import { RecordedGamePlayerMetadataBooleans, RecordedGamePlayerMetadataNumbers, RecordedGamePlayerMetadataStrings } from "@/types/RecordedGame";
 
 const Schema = mongoose.Schema;
 
 const PlayerDataSchema = new Schema(
-  RecordedGamePlayerMetadataKeySchema,
+  {
+    ...recMetadataSchemaHelper(RecordedGamePlayerMetadataBooleans, Boolean, false),
+    ...recMetadataSchemaHelper(RecordedGamePlayerMetadataStrings, String, ""),
+    ...recMetadataSchemaHelper(RecordedGamePlayerMetadataNumbers, Number, 0),
+  },
   { timestamps: true }
 );
