+++ src/db/mongo/schemas/recordedGame/RecordedGameSchema.ts
@@ -1,6 +1,7 @@
 import mongoose from "mongoose";
 import PlayerDataSchema from "./PlayerData";
-import { RecordedGameMetadataKeySchema } from "@/types/RecordedGame";
+import { recMetadataSchemaHelper } from "@/utils/utils";
+import { RecordedGameMetadataBooleans, RecordedGameMetadataNumbers, RecordedGameMetadataStrings } from "@/types/RecordedGame";
 
 const Schema = mongoose.Schema;
 
@@ -9,7 +10,9 @@ const RecordedGameSchema = new Schema(
     playerdata: { type: [PlayerDataSchema], required: true },
     buildnumber: { type: Number, required: true, default: 0}, 
     buildstring: { type: String, required: true, default: ""}, 
-    ...RecordedGameMetadataKeySchema,
+    ...recMetadataSchemaHelper(RecordedGameMetadataBooleans, Boolean, false),
+    ...recMetadataSchemaHelper(RecordedGameMetadataStrings, String, ""),
+    ...recMetadataSchemaHelper(RecordedGameMetadataNumbers, Number, 0),
   },
   { timestamps: true }
 );
